Add solution for level 3
Create a new request builder to handle FAQ. Each question corresponds to
an intent that you should configure and train on the SAP CAI platform.
The "codes" of the questions can be found in the Coop lib. Again, there
are hard-coded in the request builder, but you can put them in
a configuration file.
That request builder is registered in MessagesController to be used.

Create a new action SendCoopAnswer, which uses Coop lib to get the
answer corresponding to the intent and build the response.
To improve the action, you can handle the case when the intent is
validated by the request builder but the question doesn't exist (anymore?) in the food coop database.

Alternatively, you could use the questions table in database to store
questions and answers, and the corresponding model Question. This way,
the questions answered by the bot can be processed differently than,
let's say, the questions on the website. You can authorize admin users
to update the questions on the back-office, using ActiveAdmin.