Add solution for level 4
Create a new request builder to handle schedule display. The user should
be authenticated to see its schedule: the action SendLogin is triggered
if the user is unknown.
That request builder is registered in MessagesController to be used.

Create a new action SendSchedule, which uses Coop lib to get the (fake)
schedule of the user and build the response.